+++ libatalk/cnid/tdb/cnid_tdb_get.c
@@ -1,5 +1,5 @@
 /*
- * $Id: cnid_tdb_get.c,v 1.2 2005-04-28 20:50:02 bfernhomberg Exp $
+ * $Id: cnid_tdb_get.c,v 1.3 2005-05-03 14:55:15 didg Exp $
  */
 
 #ifdef HAVE_CONFIG_H
@@ -11,7 +11,7 @@
 #include "cnid_tdb.h"
 
 /* Return CNID for a given did/name. */
-cnid_t cnid_tdb_get(struct _cnid_db *cdb, const cnid_t did, char *name, const int len)
+cnid_t cnid_tdb_get(struct _cnid_db *cdb, const cnid_t did, char *name, const size_t len)
 {
     char start[TDB_DID_LEN + MAXPATHLEN + 1], *buf;
     struct _cnid_tdb_private *db;
